WIP refactor rename adrepository and create interface
