+++ marketplace/internal/pkg/domain/ad/repository.go
@@ -1,0 +1,7 @@
+package ad
+
+type Repository interface {
+	SaveAd(ad Ad) Ad
+	FindAdById(id string) (Ad, error)
+	FindAllAds() (adResponse []Ad)
+}
